rename configs of mixture

+++ train/myppo/env_debug.py
@@ -42,7 +42,8 @@ def env_debug(flow_params=None):
     device = torch.device("cuda:0" if args.cuda else "cpu")
 
     save_path = os.path.join(os.path.join(args.save_dir, args.algo), 'debug')
-    shutil.rmtree(save_path)
+    if os.path.exists(save_path):
+        shutil.rmtree(save_path)
     os.makedirs(save_path, exist_ok=True)
     # actor_critic, ob_rms = torch.load(pt, map_location='cpu')
